+++ test/libsolidity/smtCheckerTests/types/bytes_length.sol
@@ -1,0 +1,13 @@
+pragma experimental SMTChecker;
+
+contract C {
+	function f() public pure {
+		bytes memory x = hex"0123";
+		assert(x.length == 2);
+	}
+	function g() public pure {
+		bytes memory x = bytes(hex"0123");
+		assert(x.length == 2);
+	}
+}
+// ----